# Microoptimize new_window_link_to implementation