--- a/app/helpers/link_helper.rb
+++ b/app/helpers/link_helper.rb
@@ -1,35 +1,15 @@
 module LinkHelper
   EXTERNAL_LINK_CLASS = 'usa-link--external'.freeze
 
-  def new_window_link_to(name = nil, options = nil, html_options = nil, &block)
-    if block
-      html_options = options
-      options = name
-      name = block
-    end
-    options ||= {}
-    html_options ||= {}
+  def new_window_link_to(name = nil, options = nil, html_options = {}, &block)
+    html_options, options, name = options, name, yield(block) if block
 
-    url = url_for(options)
-    html_options[:href] ||= url
-    html_options[:class] ||= html_options[:class].to_s
     html_options[:target] = '_blank'
+    html_options[:class] = [*html_options[:class], EXTERNAL_LINK_CLASS]
 
-    classes = html_options[:class].split(' ').append(EXTERNAL_LINK_CLASS)
+    name = safe_join([name, content_tag('span', t('links.new_window'), class: 'usa-sr-only')], ' ')
 
-    html_options[:class] = classes.uniq.join(' ')
-
-    if block
-      link_to(url, html_options) do
-        yield(block)
-        concat content_tag('span', t('links.new_window'), class: 'usa-sr-only')
-      end
-    else
-      link_to(url, html_options) do
-        content_tag('span', name) +
-          content_tag('span', t('links.new_window'), class: 'usa-sr-only')
-      end
-    end
+    link_to(name, options, html_options)
   end
 
   def button_or_link_to(name = nil, options = nil, html_options = nil, &block)