--- a/app/views/shared/_footer_lite.html.erb
+++ b/app/views/shared/_footer_lite.html.erb
@@ -16,9 +16,7 @@
                                                       'aria-label': t('shared.footer_lite.gsa') } do %>
             <%= image_tag asset_url('sp-logos/square-gsa.svg'),
                           size: 20, class: 'float-left margin-right-1', alt: '' %>
-            <span>
-              <%= t('shared.footer_lite.gsa') %>
-            </span>
+            <%= t('shared.footer_lite.gsa') %>
           <% end %>
         </div>
       </div>