## 🛠 Summary of changes

Simplifies and optimizes the implementation of `new_window_link_to`, which had been showing as consuming some time on flamegraphs for profiling critical flows.

Side benefits:

- Simpler
- More compatible with `link_to` signature (e.g. support for `class` as an array of strings)
- Adds a space between text and accessible label ("Link(Opens in new window)" ➡️  "Link (Opens in new window)")
- Includes missing test coverage

Benchmark:

```rb
require 'benchmark/ips'

ENV['RAILS_ENV'] = 'production'
require File.expand_path('./config/environment.rb', __dir__)

class BenchmarksController < ActionController::Base
  include LinkHelper
end
controller_view = BenchmarksController.new.view_context

Benchmark.ips do |x|
  x.time = 10
  x.warmup = 2

  x.report('before') { controller_view.old_new_window_link_to('Link', '') }
  x.report('after') { controller_view.new_window_link_to('Link', '') }

  x.compare!
end
```

Result:

tl;dr: ~15% improvement

```
$ ruby benchmark.rb
Warming up --------------------------------------
              before     1.503k i/100ms
               after     1.696k i/100ms
Calculating -------------------------------------
              before     15.084k (± 7.9%) i/s -    150.300k in  10.077435s
               after     17.419k (± 2.6%) i/s -    174.688k in  10.035779s

Comparison:
               after:    17419.4 i/s
              before:    15084.2 i/s - 1.15x  (± 0.00) slower
```

## 📜 Testing Plan

- Spot check "Opens in new tab" links (e.g. footer links)
- `rspec spec/helpers/link_helper_spec.rb`